requirements_common.txt
ansible
celery
django
djangorestframework
django-cors-headers
django-model-utils
django-uuidfield
factory_boy
